new signature for sequencer token and set the balance threshold

+++ internal/computing/cron_task.go
@@ -31,8 +31,8 @@ func NewCronTask(nodeId string) *CronTask {
 }
 
 func (task *CronTask) RunTask() {
-	addNodeLabel()
 	checkJobStatus()
+	task.addLabelToNode()
 	task.checkCollateralBalance()
 	task.cleanAbnormalDeployment()
 	task.setFailedUbiTaskStatus()
@@ -59,6 +59,27 @@ func checkJobStatus() {
 					}
 					return true
 				})
+			}
+		}
+	}()
+}
+
+func (task *CronTask) addLabelToNode() {
+	ticker := time.NewTicker(10 * time.Minute)
+	defer ticker.Stop()
+
+	go func() {
+		for {
+			select {
+			case <-ticker.C:
+				func() {
+					defer func() {
+						if err := recover(); err != nil {
+							logs.GetLogger().Errorf("failed to add label for cluster node, error: %+v", err)
+						}
+					}()
+					addNodeLabel()
+				}()
 			}
 		}
 	}()
@@ -159,7 +180,7 @@ func (task *CronTask) watchExpiredTask() {
 		if len(deployOnK8s) == 0 && len(jobList) > 0 {
 			for _, job := range jobList {
 				if job.CreateTime > 0 {
-					if time.Now().Sub(time.Unix(job.CreateTime, 0)) > time.Hour {
+					if time.Now().Sub(time.Unix(job.CreateTime, 0)) > 2*time.Hour && job.K8sDeployName == "" {
 						if err = NewJobService().DeleteJobEntityBySpaceUuId(job.SpaceUuid, models.JOB_COMPLETED_STATUS); err != nil {
 							logs.GetLogger().Infof("failed to delete job from db, space_uuid: %s, error: %v", job.SpaceUuid, err)
 						}
